fix repo structure
