+++ backend/templates/index.html
@@ -7,11 +7,21 @@
     <link rel="stylesheet" href="{{url_for('static', filename='style.css')}}"/>
 </head>
 <body>
-    <h1>Do you allow us to save your data?</h1>
     <div class="content">
-        <a href="/form"><button>Yes</button></a>
-
-        <form method="post"><button type="submit">No</button></form>
+        <form method="POST">
+            {% if not client_data %}
+            <h1>Do you allow us to save your data?</h1>
+            <p><button type="yes" name="action" value="yes">Yes</button></p>
+            <p><button type="no" name="action" value="no">No</button></p>
+            
+            {% else %}
+            <p>Your record is already exist: <b>{{ client_data }}</b></p>
+            <p>Your can update your data or delete record </p>
+            <p><button type="update" name="action" value="update">Update</button></p>
+            <p><button type="delete" name="action" value="delete">Delete</button></p>
+            
+            {% endif %}
+        </form>
     </div>
 </body>
 </html>
