--- templates/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>MongDB Flask App</title>
-    <link rel="stylesheet" href="{{url_for('static', filename='style.css')}}"/>
-</head>
-<body>
-    <div class="content">
-        <form method="POST">
-            {% if not client_data %}
-            <h1>Do you allow us to save your data?</h1>
-            <p><button type="yes" name="action" value="yes">Yes</button></p>
-            <p><button type="no" name="action" value="no">No</button></p>
-            
-            {% else %}
-            <p>Your record is already exist: <b>{{ client_data }}</b></p>
-            <p>Your can update your data or delete record </p>
-            <p><button type="update" name="action" value="update">Update</button></p>
-            <p><button type="delete" name="action" value="delete">Delete</button></p>
-            
-            {% endif %}
-        </form>
-    </div>
-</body>
-</html>
